flytt @types/person og mer til api-modul
