+++ src/global.d.ts
@@ -23,7 +23,6 @@ declare module "Domene" {
   export type GeneriskAdresse = import("./@types").GeneriskAdresse;
   export type MedlPeriode = import("./@types").MedlPeriode;
   export type Periode = import("./@types").Periode;
-  export type Person = import("./@types").Person;
   export type OppsummertFaktaVirksomheter = import("./@types").Virksomheter;
   export type OppsummertFaktaMedfolgendeFamilie = import("./@types").MedfolgendeFamilie;
   export type MedfolgendeFamiliemedlem = import("./@types").MedfolgendeFamiliemedlem;
@@ -36,7 +35,6 @@ declare module "Domene" {
   export type SedPdfData = import("./@types").SedPdfData;
   export type Medlemskapsperiode = import("./@types").Medlemskapsperiode;
   export type OppdaterMedlemskapsperiode = import("./@types").OppdaterMedlemskapsperiode;
-  export type Familiemedlem = import("./@types").Familiemedlem;
   export type Avgiftsgrunnlag = import("./@types").Avgiftsgrunnlag;
   export type Avgiftsberegning = import("./@types").Avgiftsberegning;
   export type Avgiftsperiode = import("./@types").Avgiftsperiode;
