+++ src/services/modules/types/index.ts
@@ -1,4 +1,14 @@
 import { RegisterAdresse, UstrukturertAdresse, MidlertidigAdresse } from "./adresse";
 import Periode from "./periode";
+import Personhistorikk from "./personhistorikk";
+import Person, { Familiemedlem } from "./person";
 
-export type { RegisterAdresse, UstrukturertAdresse, MidlertidigAdresse, Periode };
+export type {
+  RegisterAdresse,
+  UstrukturertAdresse,
+  MidlertidigAdresse,
+  Periode,
+  Personhistorikk,
+  Person,
+  Familiemedlem,
+};
