+++ src/services/modules/types/person.ts
@@ -1,0 +1,27 @@
+import { KTObject } from "@navikt/melosys-kodeverk";
+
+export interface Familiemedlem {
+  fnr: string;
+  sammensattNavn: string;
+  relasjonstype: KTObject;
+  alder: number | null;
+  borMedBruker: boolean;
+  sivilstand: KTObject | null;
+  sivilstandGyldighetsperiodeFom: string | null;
+  fnrAnnenForelder: string | null;
+}
+
+interface Person {
+  erEgenAnsatt: boolean;
+  personStatus: KTObject;
+  sammensattNavn: string;
+  kjoenn: KTObject;
+  fnr: string;
+  foedselsdato: string;
+  sivilstand: KTObject;
+  statsborgerskap: KTObject;
+  statsborgerskapDato: string;
+  familiemedlemmer: Familiemedlem[];
+}
+
+export default Person;
