+++ src/services/modules/types/personhistorikk.ts
@@ -1,0 +1,19 @@
+import Periode from "./periode";
+import { RegisterAdresse, MidlertidigAdresse, UstrukturertAdresse } from "./adresse";
+
+interface Personhistorikk {
+  bostedsadressePerioder: {
+    bostedsadresse: RegisterAdresse;
+    periode: Periode;
+  }[];
+  postadressePerioder: {
+    postadresse: UstrukturertAdresse;
+    periode: Periode;
+  }[];
+  midlertidigAdressePerioder: {
+    midlertidigAdresse: MidlertidigAdresse;
+    periode: Periode;
+  }[];
+}
+
+export default Personhistorikk;
